requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
Pillow
torch
torchvision
opencv-python
mediapipe
ultralytics
transformers
timm
faiss-cpu
psycopg[binary]
pgvector
onnx
onnxruntime
onnx-simplifier
onnxscript
